Nasty hack to work around '=' in http params
In our application we got HTTP requests that resembled the following:
http://127.0.0.1:8080/path?PARAM1=This%20is%20a%20test=Something&PARAM2=1

Which contains an unescaped '=' in the value of a param. ghtp_utils:parse_query/2 errors with a case clause because it never expects this to happen.

I tried the standard Python library to parse the same URL:

>>> from urlparse import urlparse, parse_qs
>>> o = urlparse('http://127.0.0.1:8080/path?PARAM1=This%20is%20a%20test=Something&PARAM2=1')
>>> o.query
'PARAM1=This%20is%20a%20test=Something&PARAM2=1'
>>> parse_qs(o.query)
{'PARAM2': ['1'], 'PARAM1': ['This is a test=Something']}

As you can see from the above - the Python query parsing library ignores the '=' because it's obviously not a key/value pair. I tried to make ghtp_utils:parse_query/2 act in the same way, since it seems like sane behaviour.